move MissingSubprojectError declaration to exceptions submodule
